--- include/ciphers.hpp
@@ -1,13 +1,0 @@
-
-/**
- * @brief Abstract base class for encoding and decoding cyphers.
- */
-
-#ifndef CYPHER_H
-#define CYPHER_H
-
-#include <string>
-#include <morse.hpp>
-#include <caeser.hpp>
-
-#endif // CYPHER_H
